Comments on Screen Class Moved UserTimer to Graphics package
